Move "*_operand" methods into ExpressionEvaluator since "missing operand" errors are specific to evaluation.
